Cleanup, moving some QML objects to OffscreenUI
