+++ libraries/ui/src/OffscreenUi.h
@@ -25,7 +25,7 @@ class OffscreenUi : public OffscreenQmlSurface, public Dependency {
 
 public:
     OffscreenUi();
-    virtual ~OffscreenUi();
+    virtual void create(QOpenGLContext* context) override;
     void show(const QUrl& url, const QString& name, std::function<void(QQmlContext*, QObject*)> f = [](QQmlContext*, QObject*) {});
     void toggle(const QUrl& url, const QString& name, std::function<void(QQmlContext*, QObject*)> f = [](QQmlContext*, QObject*) {});
     bool shouldSwallowShortcut(QEvent* event);
